requirements.txt
amqp==2.3.2
beautifulsoup4==4.7.1
billiard==3.5.0.4
celery==4.2.1
codegen==1.0
cycler==0.10.0
Django==2.0.7
django-filter==2.0.0
djangorestframework==3.9.1
h5py==2.8.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
kombu==4.2.1
Markdown==3.0.1
matplotlib==3.0.2
numpy==1.14.5
opencv-python==3.4.1.15
patterns==0.3
Pillow==5.2.0
postgres==2.2.2
protobuf==3.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.5
PyYAML==3.13
scipy==1.1.0
six==1.11.0
soupsieve==1.7.3
tensorflow==0.12.0
vine==1.1.4
